pyproject.toml
[project]
name = "snt-kan-cnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastkan",
    "hydra-core>=1.3.2",
    "lightning>=2.5.5",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pykan>=0.2.8",
    "scikit-learn>=1.7.2",
    "tensorboard>=2.20.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]


[tool.uv.sources]
fastkan = { path = "fast-kan" }

[project.scripts]
train = "src.scripts.train:train"
